requirements.txt
streamlit==1.32.0
transformers==4.38.2
torch==2.2.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
scikit-learn==1.4.0
numpy==1.26.4
peft==0.9.0 